move funnel bar chart to own folder
