+++ frontend/src/scenes/funnels/FunnelBarChart/FunnelBarChart.scss
@@ -1,4 +1,4 @@
-@import '../../styles/mixins';
+@import '../../../styles/mixins';
 
 .FunnelBarChart {
     position: relative;
