+++ frontend/src/scenes/funnels/FunnelBarChart/FunnelBarChart.tsx
@@ -1,6 +1,6 @@
 import { useActions, useValues } from 'kea'
 import { useMemo, useRef } from 'react'
-import { funnelLogic } from './funnelLogic'
+import { funnelLogic } from '../funnelLogic'
 import './FunnelBarChart.scss'
 import { AvailableFeature, ChartParams, FunnelStepWithConversionMetrics } from '~/types'
 import { LemonRow } from 'lib/lemon-ui/LemonRow'
@@ -9,13 +9,13 @@ import { EntityFilterInfo } from 'lib/components/EntityFilterInfo'
 import { getActionFilterFromFunnelStep } from 'scenes/insights/views/Funnels/funnelStepTableUtils'
 import { IconSchedule, IconTrendingFlat, IconTrendingFlatDown } from 'lib/lemon-ui/icons'
 import { capitalizeFirstLetter, humanFriendlyDuration, percentage, pluralize } from 'lib/utils'
-import { ValueInspectorButton } from './ValueInspectorButton'
+import { ValueInspectorButton } from '../ValueInspectorButton'
 import clsx from 'clsx'
 import { useScrollable } from 'lib/hooks/useScrollable'
 import { useResizeObserver } from 'lib/hooks/useResizeObserver'
 import { getSeriesColor } from 'lib/colors'
-import { useFunnelTooltip } from './useFunnelTooltip'
-import { FunnelStepMore } from './FunnelStepMore'
+import { useFunnelTooltip } from '../useFunnelTooltip'
+import { FunnelStepMore } from '../FunnelStepMore'
 import { userLogic } from 'scenes/userLogic'
 
 function StepBarLabels(): JSX.Element {
